Add HeroComponent View Component
