+++ decidim-core/app/components/decidim/base_component.rb
@@ -1,0 +1,15 @@
+# frozen_string_literal: true
+
+module Decidim
+  class BaseComponent < ViewComponent::Base
+    delegate :current_component,
+             :current_organization,
+             :current_participatory_space,
+             :current_user,
+             :translated_attribute,
+             :icon,
+             :decidim_sanitize_admin,
+             :decidim,
+             to: :helpers
+  end
+end
