+++ decidim-core/app/components/decidim/content_blocks/hero_component/cta_button_component.html.erb
