+++ decidim-core/app/components/decidim/content_blocks/hero_component/hero_component.html.erb
@@ -8,6 +8,6 @@
       <% end %>
     </h1>
 
-    <%= render :cta_button %>
+    <%= cta_button %>
   </div>
 </section>
