+++ decidim-core/app/views/decidim/homepage/show.html.erb
@@ -11,6 +11,10 @@
   <% Decidim::ContentBlock.published.for_scope(:homepage, organization: current_organization).each do |content_block| %>
     <% next unless content_block.manifest %>
 
-    <%= cell content_block.manifest.cell, content_block %>
+    <% if content_block.manifest.component %>
+      <%= render("#{content_block.manifest.cell}_component".classify.constantize.new(content_block)) %>
+    <% else %>
+      <%= cell content_block.manifest.cell, content_block %>
+    <% end %>
   <% end %>
 </main>
